more refactoring to clean up structure
