--- src/main/scala/com/kaizen/api/Movie.scala
@@ -1,13 +1,0 @@
-package com.kaizen.api
-
-import com.kaizen.api.movie.{MovieData, MovieId, MovieTitle}
-
-final case class Movie(id: MovieId, title: MovieTitle)
-
-object Movie {
-  def fromMovieData(data: MovieData): Movie =
-    Movie(
-      data.id,
-      data.title
-    )
-}
