--- src/main/scala/com/kaizen/api/PageParams.scala
@@ -1,3 +1,0 @@
-package com.kaizen.api
-
-final case class PageParams(count: Long, before: Option[String], after: Option[String])
