--- src/main/scala/com/kaizen/api/actedIn/ActedInData.scala
@@ -1,6 +1,0 @@
-package com.kaizen.api.actedIn
-
-import com.kaizen.api.actor.ActorId
-import com.kaizen.api.movie.MovieId
-
-final case class ActedInData(actorId: ActorId, movieId: MovieId)
