--- src/main/scala/com/kaizen/api/actor/ActorData.scala
@@ -1,3 +1,0 @@
-package com.kaizen.api.actor
-
-final case class ActorData(id: ActorId, name: ActorName)
