--- src/main/scala/com/kaizen/api/actor/package.scala
@@ -1,6 +1,0 @@
-package com.kaizen.api
-
-package object actor {
-  type ActorId   = Long
-  type ActorName = String
-}
