+++ src/main/scala/com/kaizen/api/graphql/PageParams.scala
@@ -1,0 +1,3 @@
+package com.kaizen.api.graphql
+
+final case class PageParams(count: Long, before: Option[String], after: Option[String])
