+++ src/main/scala/com/kaizen/api/internal/RepositoryError.scala
@@ -1,0 +1,7 @@
+package com.kaizen.api.internal
+
+sealed trait RepositoryError extends Throwable with Product with Serializable
+
+object RepositoryError {
+  final case class ItemNotFound[ID](itemId: ID) extends RepositoryError
+}
