+++ src/main/scala/com/kaizen/api/internal/actor/ActorData.scala
@@ -1,0 +1,3 @@
+package com.kaizen.api.internal.actor
+
+final case class ActorData(id: ActorId, name: ActorName)
