+++ src/main/scala/com/kaizen/api/internal/actor/package.scala
@@ -1,0 +1,6 @@
+package com.kaizen.api.internal
+
+package object actor {
+  type ActorId   = Long
+  type ActorName = String
+}
