+++ src/main/scala/com/kaizen/api/internal/movie/MovieData.scala
@@ -1,0 +1,3 @@
+package com.kaizen.api.internal.movie
+
+final case class MovieData(id: MovieId, title: MovieTitle)
