+++ src/main/scala/com/kaizen/api/internal/movie/package.scala
@@ -1,0 +1,6 @@
+package com.kaizen.api.internal
+
+package object movie {
+  type MovieId    = Long
+  type MovieTitle = String
+}
