--- src/main/scala/com/kaizen/api/movie/MovieData.scala
@@ -1,3 +1,0 @@
-package com.kaizen.api.movie
-
-final case class MovieData(id: MovieId, title: MovieTitle)
