--- src/main/scala/com/kaizen/api/movie/package.scala
@@ -1,6 +1,0 @@
-package com.kaizen.api
-
-package object movie {
-  type MovieId    = Long
-  type MovieTitle = String
-}
